+++ frontend/src/modules/app/AppConstants.ts
