+++ frontend/src/modules/event-logs/EventLogsModal.tsx
